requirements.txt
python-dotenv
langchain
langchain-google-genai
PyPDF2
faiss-cpu
streamlit
pypdf
langchain-community
-e .